+++ R/removeArea.R
@@ -222,7 +222,6 @@ checkRemovedArea <- function(area, all_files = TRUE, opts = antaresRead::simOpti
   
   bc_not_remove <- union(bc_not_remove_cluster, bc_not_remove_link)
   if (!identical(bc_not_remove, character(0))) {
-    message("The following binding constraints have the area to remove in a coefficient : ", paste0(bc_not_remove, collapse = ", "))
-    stop("Can not remove the area ", name)
+    warning("The following binding constraints have the area to remove in a coefficient : ", paste0(bc_not_remove, collapse = ", "))
   }
 }